Remove Junit dependency; move test
