+++ spi/common/core-spi/build.gradle.kts
@@ -28,8 +28,6 @@ dependencies {
 
     testImplementation(project(":core:common:junit"))
 
-    // needed by the abstract test spec located in testFixtures
-    testFixturesImplementation(project(":core:common:junit"))
     testFixturesImplementation(libs.bundles.jupiter)
     testFixturesImplementation(libs.mockito.core)
     testFixturesImplementation(libs.assertj)
